+++ database/migrations/2017_07_30_135413_create_charges_table.php
@@ -1,22 +1,23 @@
 <?php
 /**
  * Jano Ticketing System
- * Copyright (C) 2016-2017 Andrew Ying
+ * Copyright (C) 2016-2019 Andrew Ying and other contributors.
  *
  * This file is part of Jano Ticketing System.
  *
  * Jano Ticketing System is free software: you can redistribute it and/or
- * modify it under the terms of the GNU General Public License v3.0 as
- * published by the Free Software Foundation. You must preserve all legal
- * notices and author attributions present.
+ * modify it under the terms of the GNU Affero General Public License
+ * v3.0 supplemented by additional permissions and terms as published at
+ * COPYING.md.
  *
  * Jano Ticketing System is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public
+ * License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 
 use Illuminate\Support\Facades\Schema;
@@ -33,8 +34,8 @@ class CreateChargesTable extends Migration
     public function up()
     {
         Schema::create('charges', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('account_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('account_id')->unsigned();
             $table->foreign('account_id')->references('id')->on('accounts');
             $table->string('description');
             $table->integer('amount');
